Move femparproj dirichlet tests to the main unit test for fem_parproj
